+++ Imperium/src/API/Types/Networking/BindingRequests.cs
@@ -11,6 +11,5 @@ namespace Imperium.API.Types.Networking;
 public readonly struct BindingUpdateRequest<T>
 {
     [SerializeField] public T Payload { get; init; }
-    [SerializeField] public bool InvokeUpdate { get; init; }
-    [SerializeField] public bool InvokeServerUpdate { get; init; }
+    [SerializeField] public bool InvokePrimaryUpdate { get; init; }
 }